+++ R/plot_rich_norm.R
@@ -5,6 +5,8 @@
 #' region demarcating the \code{\link{rich.core}} cutoff.
 #'
 #' @param rich.dt A \code{data.table} with rich-club coefficients
+#' @param facet.by A character string indicating whether the variable of
+#' interest is "density" or "threshold" (e.g. with DTI data)
 #' @param densities A numeric vector of the densities to plot
 #' @param alpha The significance level (DEFAULT: 0.05)
 #' @param g A list (of lists) of \code{igraph} graph objects; required if you
@@ -19,16 +21,17 @@
 #' richPlot <- plot_rich_norm(rich.dt, densities[N:(N+1)], g=g)
 #' }
 
-plot_rich_norm <- function(rich.dt, densities, alpha=0.05, g=NULL) {
+plot_rich_norm <- function(rich.dt, facet.by=c('density', 'threshold'),
+                           densities, alpha=0.05, g=NULL) {
   p.fdr <- yloc <- Group <- phi <- xstart <- xend <- NULL
 
-  subDT <- rich.dt[density %in% densities]
+  subDT <- rich.dt[eval(parse(text=facet.by)) %in% densities]
   subDT[, star := ifelse(p.fdr < alpha, '*', '')]
   subDT[, yloc := lapply(.SD, function(x)
                          min(x, na.rm=T) - 0.05 * diff(range(x, na.rm=T))),
         by=density, .SDcols='phi']
   subDT[Group == levels(subDT$Group)[2], yloc := yloc - 0.005]
-  setkey(subDT, density, Group)
+  setkeyv(subDT, c(facet.by, 'Group'))
 
   if (!is.null(g)) {
     densities.g <- round(sapply(g[[1]], graph_attr, 'density'), 2)
@@ -59,7 +62,7 @@ plot_rich_norm <- function(rich.dt, densities, alpha=0.05, g=NULL) {
     geom_rect(data=rects,
               aes(x=NULL, y=NULL, xmin=xstart, xmax=xend, ymin=-Inf, ymax=Inf),
               alpha=0.08, fill='red') +
-    facet_wrap(~ density, scales='free') +
+    facet_wrap(as.formula(paste('~', facet.by)), scales='free') +
     xlab('Degree (k)') + ylab(expression(phi[norm])) +
     theme(legend.position=c(0, 1), legend.justification=c(0, 1),
           legend.background=element_rect(size=0.5))